settingpage design
